#rnx
#rb None

[CL 4208955 by Dan Oconnor in Dev-Framework branch]

+++ Engine/Source/Runtime/Core/Public/Containers/CircularQueue.h
@@ -47,7 +47,7 @@ public:
 	 */
 	uint32 Count() const
 	{
-		int64 Count = Tail.Load() - Head.Load();
+		int32 Count = Tail.Load() - Head.Load();
 
 		if (Count < 0)
 		{
